Organise files into appropriate packages
