+++ src/main/java/lindi/commands/CreateTaskCommand.java
@@ -1,11 +1,18 @@
+package lindi.commands;
+
+import lindi.storage.Storage;
+import lindi.task.Task;
+import lindi.task.TaskList;
+import lindi.task.CreateTaskException;
+
 /**
- * Represents a command that creates a Task upon execution
+ * Represents a command that creates a lindi.task.Task upon execution
  */
 public class CreateTaskCommand extends Command {
     private final String commandString;
 
     /**
-     * Creates command to create Task with the arguments supplied by the user
+     * Creates command to create lindi.task.Task with the arguments supplied by the user
      *
      * @param commandString user input string
      */
@@ -16,14 +23,14 @@ public class CreateTaskCommand extends Command {
     /**
      * {@inheritDoc}
      *
-     * Creates task relying on Task class factory method, adds it to task list.
+     * Creates lindi.task relying on lindi.task.Task class factory method, adds it to lindi.task list.
      */
     @Override
     public void execute(TaskList tasks, Storage storage) {
         try{
             Task createdtask = Task.create(this.commandString);
             tasks.add(createdtask);
-            this.statusMsg = String.format("Okay. I've deleted this task:\n\t%s\nNow you have %d tasks in the list.",
+            this.statusMsg = String.format("Okay. I've deleted this lindi.task:\n\t%s\nNow you have %d tasks in the list.",
                     createdtask, tasks.size());
             storage.saveToFile(tasks);
         } catch (CreateTaskException e) {
