+++ src/main/java/lindi/commands/ListCommand.java
@@ -1,0 +1,19 @@
+package lindi.commands;
+
+import lindi.storage.Storage;
+import lindi.task.TaskList;
+
+/**
+ * Represents a command that lists out tasks in the lindi.task list upon execution
+ */
+public class ListCommand extends Command {
+    /**
+     * {@inheritDoc}
+     *
+     * Loads the tasks in lindi.task list into status msg to be printed.
+     */
+    @Override
+    public void execute(TaskList tasks, Storage storage) {
+        this.statusMsg = tasks.toString();
+    }
+}
