+++ src/main/java/lindi/storage/StorageLoadException.java
@@ -1,3 +1,5 @@
+package lindi.storage;
+
 public class StorageLoadException extends Exception {
     public StorageLoadException(String s) {
         super(s);
