+++ src/main/java/lindi/task/CreateDeadlineException.java
@@ -1,3 +1,5 @@
+package lindi.task;
+
 public class CreateDeadlineException extends CreateTaskException {
     public CreateDeadlineException(String s) {
         super(s);
