+++ src/main/java/lindi/task/CreateEventException.java
@@ -1,3 +1,5 @@
+package lindi.task;
+
 public class CreateEventException extends CreateTaskException {
     public CreateEventException(String s) {
         super(s);
