+++ src/main/java/lindi/task/CreateTaskException.java
@@ -1,3 +1,5 @@
+package lindi.task;
+
 public class CreateTaskException extends Exception {
     public CreateTaskException(String s) {
         super(s);
