+++ src/main/java/lindi/task/CreateToDoException.java
@@ -1,3 +1,5 @@
+package lindi.task;
+
 public class CreateToDoException extends CreateTaskException {
     public CreateToDoException(String s) {
         super(s);
