+++ src/main/java/lindi/task/ToDo.java
@@ -1,3 +1,5 @@
+package lindi.task;
+
 public class ToDo extends Task {
     public ToDo(String description) {
         super(description);
